+++ .idea/dataSources.xml
@@ -1,23 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="DataSourceManagerImpl" format="xml" multifile-model="true">
-    <data-source source="LOCAL" name="@localhost" uuid="7766630f-bc1f-4f80-95cc-2fee857dfda3">
-      <driver-ref>mariadb</driver-ref>
-      <synchronize>true</synchronize>
-      <jdbc-driver>org.mariadb.jdbc.Driver</jdbc-driver>
-      <jdbc-url>jdbc:mariadb://localhost:3306</jdbc-url>
-    </data-source>
-    <data-source source="LOCAL" name="@localhost [2]" uuid="22d1d2b9-a20a-4a9a-ae04-4f14a3670c89">
-      <driver-ref>mariadb</driver-ref>
-      <synchronize>true</synchronize>
-      <jdbc-driver>org.mariadb.jdbc.Driver</jdbc-driver>
-      <jdbc-url>jdbc:mariadb://localhost:3306</jdbc-url>
-    </data-source>
-<<<<<<< HEAD
-    <data-source source="LOCAL" name="MariaDB - @localhost" uuid="433ef5d4-6783-4b6e-8382-3f6402678931">
-=======
-    <data-source source="LOCAL" name="MariaDB - @localhost" uuid="65c51dcd-c4bf-4993-908d-d8f8aad87e6b">
->>>>>>> cbefc4407d10e554ffd60d12da5acb9b2c8803a0
+    <data-source source="LOCAL" name="@localhost" uuid="4d2e9137-0ba8-477a-b080-591f65ada17a">
       <driver-ref>mariadb</driver-ref>
       <synchronize>true</synchronize>
       <jdbc-driver>org.mariadb.jdbc.Driver</jdbc-driver>
