+++ .idea/dataSources/4d2e9137-0ba8-477a-b080-591f65ada17a/storage_v2/_src_/schema/information_schema.FNRwLQ.meta
@@ -1,0 +1,2 @@
+#n:information_schema
+!<md> [null, 0, null, null, -2147483648, -2147483648]
