+++ src/Controller/ProfilSortiController.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace App\Controller;
+
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\Routing\Annotation\Route;
+
+class ProfilSortiController extends AbstractController
+{
+    /**
+     * @Route("admin/promo/{id}/profilsortie/{num}", name="supprimer_user")
+     */
+   public function recupeApprenant(){
+
+   }
+}
